gen_provides.sh: Exclude development/mocaccino-overlay-x for provides

+++ scripts/gen_provides.sh
@@ -6,15 +6,26 @@ packages=()
 IMAGES_DATA=$(luet tree images -t $COMMON_TREE --image-repository $REPO_CACHE -t $TREE $PACK -o json)
 PKG_LIST=$(luet tree pkglist --tree $TREE --tree=$COMMON_TREE -o json)
 images=($(echo $IMAGES_DATA | jq -r '.packages[].image' ))
+images_names=($(echo $IMAGES_DATA | jq -r '.packages[].name' ))
+
+echo "For layer $PACK:
+previous packages layer: ${images_names[-2]}
+after packages layer: ${images_names[-1]}
+"
 
 if [[ "${#images[@]}" == "1" ]];then
-packages_before=()
+  packages_before=()
 else
-packages_before=( $(docker run --rm ${images[-2]} qlist -ISC) )
+  if [ "${images_names[-2]}" == "development/mocaccino-overlay-x" ] ; then
+    packages_before=()
+  else
+    packages_before=( $(docker run --rm ${images[-2]} qlist -ISC) )
+  fi
 fi
 packages_after=( $(docker run --rm ${images[-1]} qlist -ICS) )
 
 IFS=/ read -a p <<< $PACK
+
 
 path=$(echo "$PKG_LIST" | jq -r ".packages[] | select(.name==\"${p[1]}\" and .category==\"${p[0]}\").path")
 for target in "${packages_before[@]}"; do
